backend/hello-I-am-api-server/hello_i_am/requirements.txt
absl-py==0.12.0
asgiref==3.5.0
astor==0.8.1
astunparse==1.6.3
attrs==21.4.0
beautifulsoup4==4.6.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
colorama==0.4.4
cycler==0.11.0
Cython==0.29.23
dill==0.3.4
Django==3.2.12
djangorestframework==3.13.1
et-xmlfile==1.1.0
Flask==2.0.2
flatbuffers==2.0
fonttools==4.28.2
future==0.18.2
gast==0.2.2
gensim==4.1.2
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.54.0
grpcio==1.42.0
h5py==2.10.0
idna==3.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
JPype1==0.7.5
keras==2.7.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
konlpy==0.5.2
libclang==12.0.0
lxml==4.6.4
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.0
numpy==1.21.4
oauthlib==3.1.1
openpyxl==3.0.9
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.4
Pillow==8.4.0
promise==2.3
protobuf==3.19.1
py4j==0.10.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyKomoran==0.1.6.post1
PyMySQL==1.0.2
pyparsing==3.0.6
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scikit-learn==1.0.1
scipy==1.4.1
seqeval==1.2.2
setuptools-scm==6.3.2
six==1.16.0
sklearn==0.0
smart-open==5.2.1
sqlparse==0.4.2
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.0
tensorflow-datasets==4.4.0
tensorflow-estimator==2.7.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.23.1
tensorflow-metadata==1.5.0
tensorflow-text==2.7.3
termcolor==1.1.0
threadpoolctl==3.0.0
tomli==1.2.2
torch==1.10.1
tqdm==4.62.3
tweepy==3.10.0
typing_extensions==4.0.0
urllib3==1.26.7
Werkzeug==2.0.2
wincertstore==0.2
wrapt==1.13.3
xlrd==2.0.1
zipp==3.6.0
